% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_manip_functions.R
\name{get_poly_extremes}
\alias{get_poly_extremes}
\title{Get Extremes of a Polygon's Coordinates}
\usage{
get_poly_extremes(sf_poly)
}
\arguments{
\item{sf_poly}{\code{sf} object representing the polygon.}
}
\value{
A \code{data.frame} with columns \code{x_min}, \code{x_max}, \code{y_min}, and \code{y_max}.
}
\description{
Identifies the outermost coordinates (min/max x and y) of a polygon.
}
\examples{
\dontrun{
extremes <- get_poly_extremes(polygon_data)
}
}
