% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_manip_functions.R
\name{is_within_dist}
\alias{is_within_dist}
\title{Identify Objects Within a Distance}
\usage{
is_within_dist(sf_obj, connect_dist)
}
\arguments{
\item{sf_obj}{\code{sf} object containing the spatial data.}

\item{connect_dist}{\code{numeric} Distance threshold (in spatial units) to identify connections.}
}
\value{
A character vector where each element is an identifier for groups of objects within the specified distance.
}
\description{
Determines which objects in an \code{sf} object are within a specified distance of each other.
}
\examples{
\dontrun{
within_dist_ids <- is_within_dist(spatial_data, 200)
}
}
